Reorganized into Hex Archetecture
