--- src/sanguinehaze/charactercreator/data/SourceData.java
@@ -1,4 +1,0 @@
-package sanguinehaze.charactercreator.data;
-
-public class SourceData {
-}
