--- src/sanguinehaze/charactercreator/data/dtos/Sex.java
@@ -1,6 +1,0 @@
-package sanguinehaze.charactercreator.data.dtos;
-
-public enum Sex {
-    Male,
-    Female
-}
