+++ src/sanguinehaze/charactercreator/domain/InvalidArgumentException.java
@@ -1,4 +1,4 @@
-package sanguinehaze.charactercreator.data.exceptions;
+package sanguinehaze.charactercreator.domain;
 
 public class InvalidArgumentException extends RuntimeException {
     public InvalidArgumentException(String message) {
