+++ src/sanguinehaze/charactercreator/domain/SourceData.java
@@ -1,0 +1,4 @@
+package sanguinehaze.charactercreator.domain;
+
+public class SourceData {
+}
