+++ src/sanguinehaze/charactercreator/domain/dtos/Sex.java
@@ -1,0 +1,6 @@
+package sanguinehaze.charactercreator.domain.dtos;
+
+public enum Sex {
+    Male,
+    Female
+}
